Chore: extract chat and chatroom domains into separate packages
