+++ src/main/java/org/example/plzdrawing/api/chat/dto/response/ResponseChat.java
@@ -2,7 +2,7 @@ package org.example.plzdrawing.api.chat.dto.response;
 
 import java.sql.Timestamp;
 import lombok.AllArgsConstructor;
-import org.example.plzdrawing.domain.chat.message.MessageType;
+import org.example.plzdrawing.domain.chat.MessageType;
 
 @AllArgsConstructor
 public class ResponseChat {
