+++ src/main/java/org/example/plzdrawing/api/chat/service/ChatServiceImpl.java
@@ -6,8 +6,8 @@ import org.example.plzdrawing.api.chat.dto.request.ChatDto;
 import org.example.plzdrawing.api.chat.dto.response.ResponseChat;
 import org.example.plzdrawing.api.chat.dto.converter.ChatConverter;
 import org.example.plzdrawing.api.chatRoom.service.ChatRoomService;
-import org.example.plzdrawing.domain.chat.message.Chat;
-import org.example.plzdrawing.domain.chat.message.ChatRepository;
+import org.example.plzdrawing.domain.chat.Chat;
+import org.example.plzdrawing.domain.chat.ChatRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
