+++ src/main/java/org/example/plzdrawing/api/chatRoom/dto/converter/ChatRoomConverter.java
@@ -1,7 +1,7 @@
 package org.example.plzdrawing.api.chatRoom.dto.converter;
 
 import org.example.plzdrawing.api.chatRoom.dto.response.ResponseChatRoom;
-import org.example.plzdrawing.domain.chat.room.ChatRoom;
+import org.example.plzdrawing.domain.chatroom.ChatRoom;
 
 public class ChatRoomConverter {
     public static ResponseChatRoom fromEntity(ChatRoom chatRoom, String counterpartNickname) {
