+++ src/main/java/org/example/plzdrawing/api/chatRoom/exception/ChatRoomAlreadyExistsException.java
@@ -2,7 +2,7 @@ package org.example.plzdrawing.api.chatRoom.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.example.plzdrawing.domain.chat.room.ChatRoom;
+import org.example.plzdrawing.domain.chatroom.ChatRoom;
 
 @Getter
 @RequiredArgsConstructor
