+++ src/main/java/org/example/plzdrawing/api/chatRoom/service/ChatRoomServiceImpl.java
@@ -15,9 +15,9 @@ import org.example.plzdrawing.api.chatRoom.dto.response.ResponseChatRoom;
 import org.example.plzdrawing.api.chatRoom.exception.ChatRoomAlreadyExistsException;
 import org.example.plzdrawing.api.member.service.MemberService;
 import org.example.plzdrawing.common.exception.RestApiException;
-import org.example.plzdrawing.domain.chat.message.Chat;
-import org.example.plzdrawing.domain.chat.room.ChatRoom;
-import org.example.plzdrawing.domain.chat.room.ChatRoomRepository;
+import org.example.plzdrawing.domain.chat.Chat;
+import org.example.plzdrawing.domain.chatroom.ChatRoom;
+import org.example.plzdrawing.domain.chatroom.ChatRoomRepository;
 import org.example.plzdrawing.domain.member.Member;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
