+++ src/main/java/org/example/plzdrawing/domain/chat/MessageType.java
@@ -1,0 +1,5 @@
+package org.example.plzdrawing.domain.chat;
+
+public enum MessageType {
+    TEXT, FILE, IMAGE
+}
