--- src/main/java/org/example/plzdrawing/domain/chat/message/MessageType.java
@@ -1,5 +1,0 @@
-package org.example.plzdrawing.domain.chat.message;
-
-public enum MessageType {
-    TEXT, FILE, IMAGE
-}
