+++ src/main/java/org/example/plzdrawing/domain/chatroom/ChatRoomRepository.java
@@ -1,4 +1,4 @@
-package org.example.plzdrawing.domain.chat.room;
+package org.example.plzdrawing.domain.chatroom;
 
 import java.util.List;
 import java.util.Optional;
