+++ src/main/java/org/example/plzdrawing/util/security/interceptor/ChatRoomMembershipInterceptor.java
@@ -5,8 +5,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.example.plzdrawing.api.auth.customuser.CustomUser;
-import org.example.plzdrawing.domain.chat.room.ChatRoom;
-import org.example.plzdrawing.domain.chat.room.ChatRoomRepository;
+import org.example.plzdrawing.domain.chatroom.ChatRoom;
+import org.example.plzdrawing.domain.chatroom.ChatRoomRepository;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
